pyproject.toml
[project]
name = "market-making"
version = "0.1.0"
description = "DRL adaptive market making system for Binance"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=0.29.1",
    "stable-baselines3>=2.2.1",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "websockets>=12.0",
    "asyncpg>=0.29.0",
    "hmmlearn>=0.3.2",
    "sortedcontainers>=2.4.0",
    "scipy>=1.12.0",
    "pandas>=2.2.0",
    "prometheus-client>=0.19.0",
    "python-binance>=1.0.19",
    "structlog>=24.1.0",
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
